Revert "Move Logger interface from usecases to interfaces"
