+++ app/infrastructure/env.go
@@ -5,11 +5,11 @@ import (
 	"os"
 	"strings"
 
-	"github.com/bmf-san/go-clean-architecture-web-application-boilerplate/app/interfaces"
+	"github.com/bmf-san/go-clean-architecture-web-application-boilerplate/app/usecases"
 )
 
 // Load is load configs from a env file.
-func Load(logger interfaces.Logger) {
+func Load(logger usecases.Logger) {
 	filePath := ".env"
 
 	f, err := os.Open(filePath)
